put one-time auxiliary functions back into TestSingleGradient
